+++ app/models/scheduled_call.rb
@@ -1,4 +1,4 @@
 class ScheduledCall < ActiveRecord::Base
-  belongs_to :schedule
+  belongs_to :occurence
   validates_presence_of :call_id
 end
