+++ spec/models/occurence_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Occurence, type: :model do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
